add back image scaling for automatic1111111
