+++ scripts/twitch.sh
@@ -24,7 +24,7 @@ QUANTIZED=0
 KEEP_HISTORY=1
 ## Pipeline Settings
 DAEMON=1
-CONTINUOUS=0
+CONTINUOUS=1
 POLL_INTERVAL=1000
 PIPELINE_CONCURRENCY=2
 ASYNC_CONCURRENCY=0
