+++ src/pipeline.rs
@@ -16,8 +16,8 @@ use crate::ndi::send_images_over_ndi;
 use crate::openai_tts::tts as oai_tts;
 use crate::openai_tts::Request as OAITTSRequest;
 use crate::openai_tts::Voice as OAITTSVoice;
-use crate::stable_diffusion::{sd, SDConfig};
 use crate::sd_automatic::sd_auto;
+use crate::stable_diffusion::{sd, SDConfig};
 use crate::ApiError;
 use image::ImageBuffer;
 use image::Rgb;
